requirements.txt
aiohttp
discord.py
dislash.py
Flask
discord-together
praw
PyNaCl
youtube_dl
pafy